requirements.txt
cloudshell-core==2.2.176
cloudshell-l1-networking-core>=1.0,<1.1
